portalicious: registrations table component
AB#31335

Moving the registrations table into a separate component so that it can easily be re-used by the "do payment" flow
