+++ interfaces/Portalicious/src/app/pages/project-registrations/project-registrations.page.ts
@@ -4,7 +4,6 @@ import {
   computed,
   inject,
   input,
-  signal,
   ViewChild,
 } from '@angular/core';
 import { Router } from '@angular/router';
@@ -18,28 +17,15 @@ import { CardModule } from 'primeng/card';
 import { RegistrationStatusEnum } from '@121-service/src/registration/enum/registration-status.enum';
 import { PermissionEnum } from '@121-service/src/user/enum/permission.enum';
 
-import { AppRoutes } from '~/app.routes';
-import { getChipDataByRegistrationStatus } from '~/components/colored-chip/colored-chip.helper';
 import { PageLayoutComponent } from '~/components/page-layout/page-layout.component';
-import {
-  QueryTableColumn,
-  QueryTableColumnType,
-  QueryTableComponent,
-  QueryTableSelectionEvent,
-} from '~/components/query-table/query-table.component';
+import { RegistrationsTableComponent } from '~/components/registrations-table/registrations-table.component';
 import { ProjectApiService } from '~/domains/project/project.api.service';
-import { RegistrationApiService } from '~/domains/registration/registration.api.service';
-import { REGISTRATION_STATUS_LABELS } from '~/domains/registration/registration.helper';
-import { Registration } from '~/domains/registration/registration.model';
+import { registrationLink } from '~/domains/registration/registration.helper';
 import { ChangeStatusDialogComponent } from '~/pages/project-registrations/components/change-status-dialog/change-status-dialog.component';
 import { ExportRegistrationsComponent } from '~/pages/project-registrations/components/export-registrations/export-registrations.component';
 import { ImportRegistrationsComponent } from '~/pages/project-registrations/components/import-registrations/import-registrations.component';
 import { SendMessageDialogComponent } from '~/pages/project-registrations/components/send-message-dialog/send-message-dialog.component';
 import { AuthService } from '~/services/auth.service';
-import {
-  PaginateQuery,
-  PaginateQueryService,
-} from '~/services/paginate-query.service';
 import { ToastService } from '~/services/toast.service';
 
 @Component({
@@ -48,13 +34,13 @@ import { ToastService } from '~/services/toast.service';
   imports: [
     PageLayoutComponent,
     CardModule,
-    QueryTableComponent,
     ButtonModule,
     ButtonGroupModule,
     SendMessageDialogComponent,
     ExportRegistrationsComponent,
     ChangeStatusDialogComponent,
     ImportRegistrationsComponent,
+    RegistrationsTableComponent,
   ],
   providers: [ToastService],
   templateUrl: './project-registrations.page.html',
@@ -63,125 +49,30 @@ import { ToastService } from '~/services/toast.service';
 })
 export class ProjectRegistrationsPageComponent {
   // this is injected by the router
-  projectId = input.required<number>();
+  protected projectId = input.required<number>();
 
-  public authService = inject(AuthService);
+  private authService = inject(AuthService);
   private router = inject(Router);
-  private paginateQueryService = inject(PaginateQueryService);
-  private registrationApiService = inject(RegistrationApiService);
   private projectApiService = inject(ProjectApiService);
   private toastService = inject(ToastService);
 
-  PermissionEnum = PermissionEnum;
-
-  @ViewChild('table')
-  private table: QueryTableComponent<Registration, never>;
+  @ViewChild('registrationsTable')
+  private registrationsTable: RegistrationsTableComponent;
   @ViewChild('sendMessageDialog')
   private sendMessageDialog: SendMessageDialogComponent;
   @ViewChild('changeStatusDialog')
   private changeStatusDialog: ChangeStatusDialogComponent;
 
   RegistrationStatusEnum = RegistrationStatusEnum;
-  paginateQuery = signal<PaginateQuery | undefined>(undefined);
-  tableSelection = signal<QueryTableSelectionEvent<Registration>>([]);
-  contextMenuRegistration = signal<Registration | undefined>(undefined);
-
-  registrationsResponse = injectQuery(
-    this.registrationApiService.getManyByQuery(
-      this.projectId,
-      this.paginateQuery,
-    ),
-  );
 
   project = injectQuery(this.projectApiService.getProject(this.projectId));
-
-  registrations = computed(() => this.registrationsResponse.data()?.data ?? []);
-  totalRegistrations = computed(
-    () => this.registrationsResponse.data()?.meta.totalItems ?? 0,
-  );
-
-  columns = computed<QueryTableColumn<Registration>[]>(() => [
-    {
-      field: 'personAffectedSequence',
-      fieldForSort: 'registrationProgramId',
-      header: $localize`PA #`,
-      getCellRouterLink: (registration) =>
-        this.registrationLink(registration.id),
-    },
-    {
-      field: 'fullName',
-      header: $localize`:@@registration-full-name:Full Name`,
-      getCellRouterLink: (registration) =>
-        this.registrationLink(registration.id),
-    },
-    {
-      field: 'registrationCreated',
-      fieldForFilter: 'registrationCreatedDate',
-      header: $localize`:@@registration-created:Registration created`,
-      type: QueryTableColumnType.DATE,
-    },
-    {
-      field: 'status',
-      header: $localize`:@@registration-status:Status`,
-      type: QueryTableColumnType.MULTISELECT,
-      options: Object.entries(REGISTRATION_STATUS_LABELS).map(
-        ([value, label]) => ({
-          label,
-          value,
-        }),
-      ),
-      getCellChipData: (registration) =>
-        getChipDataByRegistrationStatus(registration.status),
-    },
-  ]);
-
-  private registrationLink = (registrationId: number) => [
-    '/',
-    AppRoutes.project,
-    this.projectId(),
-    AppRoutes.projectRegistrations,
-    registrationId,
-  ];
-
-  getActionData({
-    triggeredFromContextMenu = false,
-  }: {
-    triggeredFromContextMenu?: boolean;
-  } = {}) {
-    let selection = this.tableSelection();
-
-    if (Array.isArray(selection) && selection.length === 0) {
-      if (triggeredFromContextMenu) {
-        const contextMenuRegistration = this.contextMenuRegistration();
-        if (!contextMenuRegistration) {
-          this.toastService.showGenericError();
-          return;
-        }
-        selection = [contextMenuRegistration];
-      } else {
-        this.toastService.showToast({
-          severity: 'error',
-          detail: $localize`:@@no-registrations-selected:Select one or more registrations and try again.`,
-        });
-        return;
-      }
-    }
-
-    return this.paginateQueryService.selectionEventToActionData({
-      selection,
-      fieldForFilter: 'referenceId',
-      totalCount: this.totalRegistrations(),
-      currentPaginateQuery: this.paginateQuery(),
-      previewItemForSelectAll: this.registrations()[0],
-    });
-  }
 
   sendMessage({
     triggeredFromContextMenu = false,
   }: {
     triggeredFromContextMenu?: boolean;
   } = {}) {
-    const actionData = this.getActionData({
+    const actionData = this.registrationsTable.getActionData({
       triggeredFromContextMenu,
     });
 
@@ -199,7 +90,7 @@ export class ProjectRegistrationsPageComponent {
     status: RegistrationStatusEnum;
     triggeredFromContextMenu?: boolean;
   }) {
-    const actionData = this.getActionData({
+    const actionData = this.registrationsTable.getActionData({
       triggeredFromContextMenu,
     });
 
@@ -211,33 +102,43 @@ export class ProjectRegistrationsPageComponent {
   }
 
   onActionComplete() {
-    this.table.resetSelection();
+    this.registrationsTable.resetSelection();
   }
 
-  canChangeStatus(
-    status:
-      | RegistrationStatusEnum.declined
-      | RegistrationStatusEnum.deleted
-      | RegistrationStatusEnum.included
-      | RegistrationStatusEnum.paused
-      | RegistrationStatusEnum.validated,
-  ) {
-    const statusToPermissionMap = {
-      [RegistrationStatusEnum.validated]:
-        PermissionEnum.RegistrationStatusMarkAsValidatedUPDATE,
-      [RegistrationStatusEnum.included]:
-        PermissionEnum.RegistrationStatusIncludedUPDATE,
-      [RegistrationStatusEnum.declined]:
-        PermissionEnum.RegistrationStatusMarkAsDeclinedUPDATE,
-      [RegistrationStatusEnum.deleted]: PermissionEnum.RegistrationDELETE,
-      [RegistrationStatusEnum.paused]:
-        PermissionEnum.RegistrationStatusPausedUPDATE,
-    };
-    return this.authService.hasPermission({
-      projectId: this.projectId(),
-      requiredPermission: statusToPermissionMap[status],
-    });
-  }
+  canChangeStatus = computed(
+    () =>
+      (
+        status:
+          | RegistrationStatusEnum.declined
+          | RegistrationStatusEnum.deleted
+          | RegistrationStatusEnum.included
+          | RegistrationStatusEnum.paused
+          | RegistrationStatusEnum.validated,
+      ) => {
+        if (
+          status === RegistrationStatusEnum.validated &&
+          !this.project.data()?.validation
+        ) {
+          return false;
+        }
+
+        const statusToPermissionMap = {
+          [RegistrationStatusEnum.validated]:
+            PermissionEnum.RegistrationStatusMarkAsValidatedUPDATE,
+          [RegistrationStatusEnum.included]:
+            PermissionEnum.RegistrationStatusIncludedUPDATE,
+          [RegistrationStatusEnum.declined]:
+            PermissionEnum.RegistrationStatusMarkAsDeclinedUPDATE,
+          [RegistrationStatusEnum.deleted]: PermissionEnum.RegistrationDELETE,
+          [RegistrationStatusEnum.paused]:
+            PermissionEnum.RegistrationStatusPausedUPDATE,
+        };
+        return this.authService.hasPermission({
+          projectId: this.projectId(),
+          requiredPermission: statusToPermissionMap[status],
+        });
+      },
+  );
 
   canSendMessage = computed(() =>
     this.authService.hasPermission({
@@ -269,18 +170,24 @@ export class ProjectRegistrationsPageComponent {
         label: $localize`Go to profile`,
         icon: 'pi pi-user',
         command: () => {
-          const registration = this.contextMenuRegistration();
+          const registration =
+            this.registrationsTable.contextMenuRegistration();
           if (!registration) {
             this.toastService.showGenericError();
             return;
           }
-          void this.router.navigate(this.registrationLink(registration.id));
+          void this.router.navigate(
+            registrationLink({
+              projectId: this.projectId(),
+              registrationId: registration.id,
+            }),
+          );
         },
       },
       {
         label: $localize`Validate`,
         icon: 'pi pi-check-circle',
-        visible: this.canChangeStatus(RegistrationStatusEnum.validated),
+        visible: this.canChangeStatus()(RegistrationStatusEnum.validated),
         command: () => {
           this.changeStatus({
             status: RegistrationStatusEnum.validated,
@@ -291,7 +198,7 @@ export class ProjectRegistrationsPageComponent {
       {
         label: $localize`Include`,
         icon: 'pi pi-check',
-        visible: this.canChangeStatus(RegistrationStatusEnum.included),
+        visible: this.canChangeStatus()(RegistrationStatusEnum.included),
         command: () => {
           this.changeStatus({
             status: RegistrationStatusEnum.included,
@@ -302,7 +209,7 @@ export class ProjectRegistrationsPageComponent {
       {
         label: $localize`Decline`,
         icon: 'pi pi-times',
-        visible: this.canChangeStatus(RegistrationStatusEnum.declined),
+        visible: this.canChangeStatus()(RegistrationStatusEnum.declined),
         command: () => {
           this.changeStatus({
             status: RegistrationStatusEnum.declined,
@@ -323,7 +230,7 @@ export class ProjectRegistrationsPageComponent {
       {
         label: $localize`Delete`,
         icon: 'pi pi-trash',
-        visible: this.canChangeStatus(RegistrationStatusEnum.deleted),
+        visible: this.canChangeStatus()(RegistrationStatusEnum.deleted),
         command: () => {
           this.changeStatus({
             status: RegistrationStatusEnum.deleted,
